requirements.txt
django==4.0.5
djangorestframework==3.13.1
Pillow==9.1.1
django-cors-headers==3.13.0
django-rest-multiple-models==2.1.3